Entrance and exit animations for work profile first run.
Slide up from the bottom, fade in. Animation happens after the
screenshot entrance animation (same time the view appeared before). This
isn't meant to be a final animation, just to put something in so it's
not so jarring.

Includes some changes in preparation for the message area being able to
support multiple types of content:
1. Moved all interactions with the view to MessageContainerController,
   which will be able to handle different types of content. This isn't
   entirely generic as it does contain the view code for the work
   profile first run.
2. Moved work profile first run view code to its own XML to let
   different content swap in more easily.

The animations themselves are only called by flag-protected code, but a
couple of aspects are not fully flag-protected:
1. The message container view is now INVISIBLE instead of GONE, and is
   shifted down (y-axis) to be outside of the containing ConstraintLayout.
2. The message container is attached to a Guideline which starts off at
   the bottom of the screen. The actions container sits atop that
   guideline. No user-visible change, just expressed with a slightly
   different structure.
3. Slight increase in the bottom margin of the actions container and the
   message container, as it was being clipped by the navbar.

Test: Visual inspection of the flow
Bug: 254245929
Change-Id: Ie1b97f841bb28de52a8a9ea9ac5ed48c3f694abc