After this change, when user enters a screen, an assist structure will be generated and sent along a field classification request to pcc classification service.
When user enters a screen, a new session will be created. Inside session creation, session#updateLocked() will be
called and it checks for the flag FLAG_PCC_DETECTION. If the flag is set, it will request autofill data with a PccAssistDataReceiverImpl
instance. This instance would then send a field classification request
after it receives the asssit structure from activity manager.

This change is guarded by PCC_ENABLED flag.

Bug:b/266582009, b/261107376

Change-Id: I636cb6480fcfbe6812de347e94814ccd0a0b7b79