Introduce utility method in CarrierConfigManager to get config subset
This CL introduces a hidden utility method for system components to
get the subset of carrier config.

Many system components has been update to use the new
CarrierConfigManager API to get subset of config. But the callers
side have repetitive private methods to do the same thing. The
utility method in CarrierConfigManager will remove the duplication.

Bug: b/263267340
Test: FrameworksTelephonyTests
Change-Id: Ifbd84c67525fd9ea2925e5b735788e6c3c22dd34