Shade - Don't force auth on pulldown with sensitive notif
This was originally seen as a shortcut to seeing the full notification
content. In practice, it is preventing phone users from accessing
quick settings as not everyone is aware of distinction between pulling
down from the top vs the middle. This also makes split shade access
impossible, so show the redacted shade instead.

Fixes: 262717969
Test: atest LockscreenShadeTransitionControllerTest
Change-Id: I9504dd5769c61b96a3fe582e114f59b53866c1ff