Make active tiles always listen on creation
When a CustomTile corresponding to an active tile is created (either on
process start or on user switch), perform one round of
startListening/stopListening. This makes sure that the TileService has
an opportunity to refresh the information after a user switch without
having to request it.

Also, make `requestListeningState` called on the wrong user a noop.

Fixes: 242194868
Test: atest CustomTileTest
Test: manual using test TileService app
Change-Id: I5aea6f849e7a75ff71df8fb6c0d242785e94b8c3