Fix SearchScaffold navigation back arrow
The issue is when search mode is on, cannot click navigation back arrow
to close the search mode.

The issue is caused by when navigation back arrow is clicked, it dismiss
the input method, which will call onSearchQueryChange again to reset the
search query and re-enter the search mode.

Split the isSearchMode from searchQuery to fix this issue.

Fix: 272574692
Test: Manually with Settings
Test: Unit test
Change-Id: Id2bcc643c8c85e57a543221a6951b00c11ae955a