feat(#MagSettingsPanel)!: support WindowMagnificationSettings for fullscreen magnifier
We adjust the MagnificationModeSwitch onSwitch callback to onClick. When tapping on MagnificationModeSwitch, we changed to open WindowMagnificationSettings. Originally it would just transit to window mode directly.

According to the adjustment, we can open settings panel for fullscreen mode. We can still transit to window mode by clicking the panel size buttons. Also, we can directly change the fullscreen magnifier scale with the panel scale seek bar.

Bug: 271127581
Test: manually
      atest MagnificationModeSwitchTest
      atest ModeSwitchesControllerTest
      atest WindowMagnificationTest
      atest WindowMagnificationSettingsTest
      atest MagnificationControllerTest
Change-Id: Ibd1f914f5ba08939280be3cf9c33554ea4d520ca