Hold onto occluded change until it's committed
When we're delaying commit of setKeyguardOccluded due to a transition
taking over instead to synchronise occlude status more nicely, we hold
onto the final occlude status so we can be eventually consistent even
if the systemui restarts or gets something wrong.

The "notify" parameter in setKeyguardOccludedLw was risky because when
false it could lead to SystemUI not being informed of an update, but
the system server still considering its job done because the flag for
pending updates was cleared anyway.

This is a speculative fix to help the attached issue resolved itself
whenever it happens.

(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:578b03797c4798fc29d3297afc6660386df2afb2)

Test: atest android.server.wm.KeyguardTests
Bug: 269892931
Merged-In: I7d3ac020db34d8a15d18b25de07231bda83cbcf0
Change-Id: I7d3ac020db34d8a15d18b25de07231bda83cbcf0