Revert "Adds assets, removes SFPS indicator, updates BP controller for rear display/tent mode"
This reverts commit e69b443ceec2d84b1804b56de9083530184be4c8.

Reason for revert: This CL causes an issue where the SFPS indicator doesn't show up when expected, because it assumes the device to be in rear display mode. The fix for this ag/21801232 won't make it for the qpr-dev cut-off, so reverting this change and cherry picking it and the associated fix to udc.
Bug: 265198124
Change-Id: I5600640cbb0e3384ba46db6e9f26dc464ca444e0